Изменена файловая структура, перенесены файлы js по БЭМ
