--- src/pages-css/about.css
@@ -1,18 +1,0 @@
-@import url('../vendor/normalize.css');
-@import url('../vendor/fonts/fonts.css');
-@import url('../blocks/general/general.css');
-@import url('../blocks/header/header.css');
-@import url('../blocks/content/__title/content__title.css');
-@import url('../blocks/content/__subtitle/content__subtitle.css');
-@import url('../blocks/content/author/author.css');
-@import url('../blocks/plane-text/plane-text.css');
-@import url('../blocks/footer/footer.css');
-@import url('../blocks/content/result/__preloader/result__preloader.css');
-@import url('../blocks/content/result/__error/result__error.css');
-
-
-@import url('../blocks/about/promo/promo.css');
-@import url('../blocks/about/skills/skills.css');
-@import url('../blocks/about/comments/comments.css');
-
-
