--- src/pages-css/style.css
@@ -1,7 +1,0 @@
-@import url('../vendor/normalize.css');
-@import url('../vendor/fonts/fonts.css');
-@import url('../blocks/header/header.css');
-@import url('../blocks/content/content.css');
-@import url('../blocks/plane-text/plane-text.css');
-@import url('../blocks/footer/footer.css');
-@import url('../blocks/general/general.css');
