+++ src/pages/about/about.html
@@ -6,10 +6,10 @@
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>NewsAnalyzer</title>
     <link rel="stylesheet" href="<%=htmlWebpackPlugin.files.chunks.about.css %>">
-    <link rel="apple-touch-icon" sizes="180x180" href="<%=require('./favicon.ico').default%>">
-    <link rel="apple-touch-icon" href="<%=require('./favicon.ico').default%>">
-    <link rel="icon" type="image/x-icon" href="<%=require('./favicon.ico').default%>">
-    <link rel="mask-icon" href="<%=require('./favicon.ico').default%>">
+    <link rel="apple-touch-icon" sizes="180x180" href="<%=require('../../favicon.ico').default%>">
+    <link rel="apple-touch-icon" href="<%=require('../../favicon.ico').default%>">
+    <link rel="icon" type="image/x-icon" href="<%=require('../../favicon.ico').default%>">
+    <link rel="mask-icon" href="<%=require('../../favicon.ico').default%>">
 </head>
 <body class="general">
     <div class="page">
@@ -39,28 +39,28 @@
                 <div class="skills__items">
                     <div class="skills__element">
                         <h4 class="plane-text skills__subtitle">HTML</h4>
-                        <img src="<%=require('./images/html.png').default%>" alt="HTML">
+                        <img src="<%=require('../../images/html.png').default%>" alt="HTML">
                     </div>
                     <div class="skills__element">
                         <h4 class="plane-text skills__subtitle">CSS</h4>
-                        <img src="<%=require('./images/css.png').default%>" alt="CSS">
+                        <img src="<%=require('../../images/css.png').default%>" alt="CSS">
                     </div>
                     <div class="skills__element">
                         <h4 class="plane-text skills__subtitle">Java Script</h4>
-                        <img src="<%=require('./images/js.png').default%>" alt="Java Script">
+                        <img src="<%=require('../../images/js.png').default%>" alt="Java Script">
                     </div> 
                     <div class="skills__element">
                         <h4 class="plane-text skills__subtitle">Webpack</h4>
-                        <img src="<%=require('./images/webpack.png').default%>" alt="Webpack">
+                        <img src="<%=require('../../images/webpack.png').default%>" alt="Webpack">
                     </div>                   
                 </div>
             </section>
 
             <section class="author">
                     <picture>
-                        <source media="(max-width: 768px)" srcset="<%=require('./images/author_small.jpg').default%>">
-                        <source srcset="<%=require('./images/real_author.jpg').default%>">   
-                        <img class="author__image" alt="Фото автора" src="<%=require('./images/real_author.jpg').default%>">
+                        <source media="(max-width: 768px)" srcset="<%=require('../../images/author_small.jpg').default%>">
+                        <source srcset="<%=require('../../images/real_author.jpg').default%>">   
+                        <img class="author__image" alt="Фото автора" src="<%=require('../../images/real_author.jpg').default%>">
                     </picture>
                     <div class="author__info">
                         <h2 class="author__title">Об авторе</h2>
@@ -93,9 +93,9 @@
                 </nav>
                 <ul class="footer__icons"> 
                     <li><a href="https://github.com/Elena-FrontendDev/NewsAnalyzer">
-                        <img class="footer__icons-item" alt="GitHub" src="<%=require('./images/gh_icon.svg').default%>"></a></li>
+                        <img class="footer__icons-item" alt="GitHub" src="<%=require('../../images/gh_icon.svg').default%>"></a></li>
                         <li><a href="https://www.facebook.com/elena.kulikova.5680">
-                    <img class="footer__icons-item" alt="Facebook" src="<%=require('./images/fb_icon.svg').default%>"></a></li>    
+                    <img class="footer__icons-item" alt="Facebook" src="<%=require('../../images/fb_icon.svg').default%>"></a></li>    
                 </ul>
                 </footer>
     </div>
