+++ src/pages/main/index.html
@@ -6,10 +6,10 @@
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>NewsAnalyzer</title>
     <link rel="stylesheet" href="<%=htmlWebpackPlugin.files.chunks.main.css %>">
-    <link rel="apple-touch-icon" sizes="180x180" href="<%=require('./favicon.ico').default%>">
-    <link rel="apple-touch-icon" href="<%=require('./favicon.ico').default%>">
-    <link rel="icon" type="image/x-icon" href="<%=require('./favicon.ico').default%>">
-    <link rel="mask-icon" href="<%=require('./favicon.ico').default%>">
+    <link rel="apple-touch-icon" sizes="180x180" href="<%=require('../../favicon.ico').default%>">
+    <link rel="apple-touch-icon" href="<%=require('../../favicon.ico').default%>">
+    <link rel="icon" type="image/x-icon" href="<%=require('../../favicon.ico').default%>">
+    <link rel="mask-icon" href="<%=require('../../favicon.ico').default%>">
 
 </head>
 <body class="general">
@@ -47,11 +47,10 @@
             </section>
             
             <section class="author">
-                <!-- <img class="author__image" alt="Фото автора" src="<%=require('./images/real_author.jpg').default%>"> -->
                 <picture>
-                    <source media="(max-width: 768px)" srcset="<%=require('./images/author_small.jpg').default%>">
-                    <source srcset="<%=require('./images/real_author.jpg').default%>">   
-                    <img class="author__image" alt="Фото автора" src="<%=require('./images/real_author.jpg').default%>">
+                    <source media="(max-width: 768px)" srcset="<%=require('../../images/author_small.jpg').default%>">
+                    <source srcset="<%=require('../../images/real_author.jpg').default%>">   
+                    <img class="author__image" alt="Фото автора" src="<%=require('../../images/real_author.jpg').default%>">
                 </picture>
                 <div class="author__info">
                     <h2 class="author__title">Об авторе</h2>
@@ -74,10 +73,10 @@
             </nav>
             <ul class="footer__icons"> 
                 <li><a href="https://github.com/Elena-FrontendDev/NewsAnalyzer">
-                    <img class="footer__icons-item" alt="GitHub" src="<%=require('./images/gh_icon.svg').default%>"></a>
+                    <img class="footer__icons-item" alt="GitHub" src="<%=require('../../images/gh_icon.svg').default%>"></a>
                 </li>
                 <li><a href="https://www.facebook.com/elena.kulikova.5680">
-                    <img class="footer__icons-item" alt="Facebook" src="<%=require('./images/fb_icon.svg').default%>"></a>    
+                    <img class="footer__icons-item" alt="Facebook" src="<%=require('../../images/fb_icon.svg').default%>"></a>    
                 </li>
             </ul>
             </footer>
