+++ src/pages/statist/statist.html
@@ -6,10 +6,10 @@
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>NewsAnalyzer</title>
     <link rel="stylesheet" href="<%=htmlWebpackPlugin.files.chunks.statistics.css %>">
-    <link rel="apple-touch-icon" sizes="180x180" href="<%=require('./favicon.ico').default%>">
-    <link rel="apple-touch-icon" href="<%=require('./favicon.ico').default%>">
-    <link rel="icon" type="image/x-icon" href="<%=require('./favicon.ico').default%>">
-    <link rel="mask-icon" href="<%=require('./favicon.ico').default%>">
+    <link rel="apple-touch-icon" sizes="180x180" href="<%=require('../../favicon.ico').default%>">
+    <link rel="apple-touch-icon" href="<%=require('../../favicon.ico').default%>">
+    <link rel="icon" type="image/x-icon" href="<%=require('../../favicon.ico').default%>">
+    <link rel="mask-icon" href="<%=require('../../favicon.ico').default%>">
 </head>
 <body class="general">
         <div class="page">
@@ -76,9 +76,9 @@
                 
                 <ul class="footer__icons"> 
                     <li><a href="https://github.com/Elena-FrontendDev/NewsAnalyzer">
-                        <img class="footer__icons-item" alt="GitHub" src="<%=require('./images/gh_icon.svg').default%>"></a></li>
+                        <img class="footer__icons-item" alt="GitHub" src="<%=require('../../images/gh_icon.svg').default%>"></a></li>
                     <li><a href="https://www.facebook.com/elena.kulikova.5680">
-                        <img class="footer__icons-item" alt="Facebook" src="<%=require('./images/fb_icon.svg').default%>"></a></li>   
+                        <img class="footer__icons-item" alt="Facebook" src="<%=require('../../images/fb_icon.svg').default%>"></a></li>   
                 </ul>
                 </footer>
             </div>
