+++ src/pages/statist/statist.js
@@ -1,0 +1,9 @@
+import "../statist/statist.css";
+
+import {statistics} from '../../js/constants/constForStatistic';
+
+window.onload = function() {
+     statistics.setRequest();
+     statistics.makeYaxisForChart();
+     statistics.checkHeaders();
+    }
